requirements.txt
python-dotenv==1.1.0
requests==2.32.3
selenium==4.34.2
webdriver-manager==4.0.2
